requirements.txt
sqlalchemy
pydantic
uvicorn
fastapi
python-multipart
Pillow
pdf2image
easyocr
torch
torchvision
torchaudio
numpy
asyncmy
alembic
python-dotenv